+++ src/main/java/de/fourofakind/businesshotel/client/src/app/components/employee-management/employee-management.component.ts
@@ -7,6 +7,7 @@ import {isNumeric} from "rxjs/internal-compatibility";
 import {Alert} from "../../app.component";
 import {BookingService} from "../../services/booking/booking.service";
 import {NgForm} from "@angular/forms";
+import {RoleService} from "../../services/role/role.service";
 
 @Component({
   selector: 'app-employee-management',
@@ -17,7 +18,8 @@ export class EmployeeManagementComponent implements OnInit {
 
   constructor(private accountdetailsService: AccountdetailsService,
               private employeeService: EmployeeService,
-              private bookingService: BookingService)
+              private bookingService: BookingService,
+              private roleService: RoleService)
   {
 
   }
@@ -27,10 +29,7 @@ export class EmployeeManagementComponent implements OnInit {
 
   }
 
-  onSubmit()
-  {
-
-  }
+  private readonly department:string="employee-management";
 
   isChecked:boolean = false;
   firstName!:string;
@@ -144,29 +143,37 @@ export class EmployeeManagementComponent implements OnInit {
 
   addOrUpdateEmployeeAndDetails(addsNewEmployee:boolean, addOrUpdateEmployeeForm:NgForm)
   {
-    if(addsNewEmployee) this.addAccount(()=>this.addOrUpdateEmployee(addsNewEmployee,addOrUpdateEmployeeForm));
-    else this.addOrUpdateEmployee(addsNewEmployee,addOrUpdateEmployeeForm)
+    if (this.roleService.checkRights(this.department))
+    {
+      if(addsNewEmployee) this.addAccount(()=>this.addOrUpdateEmployee(addsNewEmployee,addOrUpdateEmployeeForm));
+      else this.addOrUpdateEmployee(addsNewEmployee,addOrUpdateEmployeeForm)
+    }
+    else alert("Benötigte Rechte nicht vorhanden")
   }
 
 
   submitSearch(intoFormular:boolean){
-    this.foundEmployee=null;
-    this.employeeService.getEmployee(this.empNo)
-      .subscribe(data=>
+    if (this.roleService.checkRights(this.department))
     {
-      this.foundEmployee=data;
-      if(intoFormular)
+      this.foundEmployee=null;
+      this.employeeService.getEmployee(this.empNo)
+        .subscribe(data=>
       {
-        this.firstName=data.empName.substring(0, data.empName.lastIndexOf(" "));
-        this.lastName=data.empName.substring(data.empName.lastIndexOf(" "));
-        this.givenRole=data.givenRole;
-        this.accountID=data.accountID;
-      }
-    },
-      (error)=>
-      {
-        this.addAlertForXSeconds(new Alert('danger',"Kein Mitarbeiter mit dieser Mitarbeiternummer vorhanden"),5);
-      })
+        this.foundEmployee=data;
+        if(intoFormular)
+        {
+          this.firstName=data.empName.substring(0, data.empName.lastIndexOf(" "));
+          this.lastName=data.empName.substring(data.empName.lastIndexOf(" "));
+          this.givenRole=data.givenRole;
+          this.accountID=data.accountID;
+        }
+      },
+        (error)=>
+        {
+          this.addAlertForXSeconds(new Alert('danger',"Kein Mitarbeiter mit dieser Mitarbeiternummer vorhanden"),5);
+        })
+    }
+    else alert("Benötigte Rechte nicht vorhanden")
   }
 
   deleteEmployee(deleteEmployeeForm: NgForm)
@@ -184,6 +191,7 @@ export class EmployeeManagementComponent implements OnInit {
           this.addAlertForXSeconds(new Alert('danger',"Kein Mitarbeiter mit dieser Mitarbeiternummer vorhanden"),5);
         }
       );
+
   }
 
   patchBookings(_callback:Function)
@@ -222,9 +230,11 @@ export class EmployeeManagementComponent implements OnInit {
   deleteEmployeeAndDetails(deleteEmployeeForm: NgForm)
   {
 
-
-    this.foundEmployee=null;
-    this.patchBookings(()=>this.deleteEmployee(deleteEmployeeForm));
-
+    if (this.roleService.checkRights(this.department))
+    {
+      this.foundEmployee=null;
+      this.patchBookings(()=>this.deleteEmployee(deleteEmployeeForm));
+    }
+    else alert("Benötigte Rechte nicht vorhanden")
   }
 }
